Move database backend selection to the build config
The `db-root` task is now called `db-config`, and now centralizes the
database connection URLs in a single file, which also contains a list
of available database backends.

The `util/dbapi` module has been adapted to allow setting of the DBAPI
root elsewhere, so it doesn't depend directly of the database config
module.  This operation is currently performed by `data/dbapi`.

Finally, it's now possible for the user to select a database backend
via a `localStorage` preference.