Get the project's version for the about modal
For this, a new multi task `build-id` was written to get git's `HEAD` id
and turn it into a SemVer-compliant version, resorting to reading
`package.json` if git isn't available.  While we were at it, the `publish`
task was rewritten to not use `grunt-gh-pages`'s internal git functions,
calling `grunt.util.spawn` directly.