Add URL fragment–based user focus
When a username is specified via a URL fragment (e.g. `#AluisioASG`),
find that user's data in the user collection and zoom the map at their
location.  This check is both done at load time (so one can link to a
particular individual) or at run time (so it can be changed without
refreshing the page).